Crate pallet_liquidation

Liquidate pallet

Overview

This pallets provides offchain worker to call the liquidate_borrow operation in loans pallet. The collator may opt-in with a pre-funded account. The liquidate strategy is:

find the unhealthy account which has exceeded loans

liquidate the currency with higher loans

liquidator gets any of the affordable collaterals.

Re-exports

pub use pallet::*;

Modules

crypto

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Constants

KEY_TYPE