Trait pallet_liquidation::pallet::Config

pub trait Config: CreateSignedTransaction<Call<Self>> + Config + Config {
    type AuthorityId: AppCrypto<Self::Public, Self::Signature>;
    type LockPeriod: Get<u64>;
    type LiquidateFactor: Get<Percent>;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Associated Types

type AuthorityId: AppCrypto<Self::Public, Self::Signature>

The account type to perform liquidation

type LockPeriod: Get<u64>

The lockdown time when running offchain worker

type LiquidateFactor: Get<Percent>

The maximum value when liquidate a loan, may different with the loans pallet.

Loading content...